requirements.txt
asgiref==3.5.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==4.0.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
filelock==3.6.0
idna==3.3
Markdown==3.3.4
MarkupPy==1.14
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.1
platformdirs==2.5.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.13.4
xlrd==2.0.1
xlwt==1.3.0
